test(api-unit-tests): change ending of test classes to "...Test"
