+++ qtaf-rest-assured/src/test/java/de/qytera/qtaf/apitesting/restassured/requestTypes/advanced/TestExecutionRequestTest.java
@@ -4,7 +4,6 @@ import de.qytera.qtaf.apitesting.ApiTest;
 import de.qytera.qtaf.apitesting.ExecutedApiTest;
 import de.qytera.qtaf.testng.context.QtafTestNGContext;
 import org.testng.Assert;
-import org.testng.annotations.Ignore;
 import org.testng.annotations.Test;
 
 import java.util.List;
@@ -12,7 +11,7 @@ import java.util.Map;
 
 import static de.qytera.qtaf.apitesting.ApiTestExecutor.apiTest;
 
-public class TestExecutionRequestTests extends QtafTestNGContext implements ApiTest {
+public class TestExecutionRequestTest extends QtafTestNGContext implements ApiTest {
     private static String url = "https://jsonplaceholder.typicode.com";
 
     @Test(testName = "Test Request Getters")
