+++ qtaf-rest-assured/src/test/java/de/qytera/qtaf/apitesting/restassured/requestTypes/basic/OptionsTest.java
@@ -3,7 +3,6 @@ package de.qytera.qtaf.apitesting.restassured.requestTypes.basic;
 import de.qytera.qtaf.apitesting.ApiTest;
 import de.qytera.qtaf.apitesting.log.model.message.ApiLogMessage;
 import de.qytera.qtaf.core.config.annotations.TestFeature;
-import de.qytera.qtaf.core.log.model.message.AssertionLogMessageType;
 import de.qytera.qtaf.core.log.model.message.LogMessage;
 import de.qytera.qtaf.testng.context.QtafTestNGContext;
 import org.testng.Assert;
@@ -16,18 +15,17 @@ import static de.qytera.qtaf.apitesting.ApiTestExecutor.apiTest;
 
 import static de.qytera.qtaf.apitesting.restassured.util.TestHelper.*;
 
-@TestFeature(name = "GET Request Tests", description = "Check the get request methods")
-public class GetTests extends QtafTestNGContext implements ApiTest {
+@TestFeature(name = "OPTIONS Request Tests", description = "Check the options request methods")
+public class OptionsTest extends QtafTestNGContext implements ApiTest {
     String url = "https://jsonplaceholder.typicode.com";
-    @Test(testName = "test getRequest() -> PASSED")
-    public void testGetRequestPASSED() {
-
+    @Test(testName = "test optionsRequest() -> PASSED")
+    public void testOptionsRequestPASSED() {
         apiTest(
                 this,
                 List.of(
-                        baseUri(url + "/users/1")
-                ),
-                getRequest(),
+                        baseUri(url),
+                        basePath("/users/1")),
+                optionsRequest(),
                 List.of()
         );
         ApiLogMessage latestApiLogMessage = getLatestApiLogMessageFromContext(this);
@@ -36,19 +34,19 @@ public class GetTests extends QtafTestNGContext implements ApiTest {
                 latestApiLogMessage,
                 LogMessage.Status.PASSED,
                 0,
-                ApiLogMessage.Action.RequestType.GET,
-                200
+                ApiLogMessage.Action.RequestType.OPTIONS,
+                204
         );
     }
 
-    @Test(testName = "test getRequest() with wrong path but no assertion -> PASSED")
-    public void testGetRequestWrongPathNoAssertionsPASSED() {
+    @Test(testName = "test optionsRequest() with wrong path but no assertion -> PASSED")
+    public void testOptionsRequestPASSEDWrongPathNoAssertionsPASSED() {
         apiTest(
                 this,
                 List.of(
-                        baseUri(url + "/user/1")
-                ),
-                getRequest(),
+                        baseUri(url),
+                        basePath("/user/1")),
+                optionsRequest(),
                 List.of()
         );
         ApiLogMessage latestApiLogMessage = getLatestApiLogMessageFromContext(this);
@@ -57,19 +55,19 @@ public class GetTests extends QtafTestNGContext implements ApiTest {
                 latestApiLogMessage,
                 LogMessage.Status.PASSED,
                 0,
-                ApiLogMessage.Action.RequestType.GET,
-                404
+                ApiLogMessage.Action.RequestType.OPTIONS,
+                204
         );
     }
 
-    @Test(testName = "test getRequest() not path -> PENDING")
-    public void testDeleteRequestNoPathPASSED() {
+    @Test(testName = "test optionsRequest() no path -> PENDING")
+    public void testOptionsRequestNoPathPASSED() {
 
         try {
             apiTest(
                     this,
                     List.of(),
-                    getRequest(),
+                    optionsRequest(),
                     List.of()
             );
         } catch (Exception e){
@@ -82,7 +80,7 @@ public class GetTests extends QtafTestNGContext implements ApiTest {
                 latestApiLogMessage,
                 LogMessage.Status.PENDING,
                 0,
-                ApiLogMessage.Action.RequestType.GET,
+                ApiLogMessage.Action.RequestType.OPTIONS,
                 0
         );
     }
